Revert "feat(runtime-vapor): extract apiSetup & init setup ctx"
This reverts commit a0b5653a5f6acff71769763564d77bc468080923.
