+++ packages/runtime-vapor/src/apiCreateComponent.ts
@@ -3,7 +3,7 @@ import {
   createComponentInstance,
   currentInstance,
 } from './component'
-import { setupComponent } from './apiSetup'
+import { setupComponent } from './apiRender'
 import type { RawProps } from './componentProps'
 
 export function createComponent(comp: Component, rawProps: RawProps = null) {
