+++ packages/runtime-vapor/src/apiRender.ts
@@ -1,6 +1,8 @@
-import type { ComponentInternalInstance } from './component'
+import { isArray, isFunction, isObject } from '@vue/shared'
+import { type ComponentInternalInstance, setCurrentInstance } from './component'
 import { insert, querySelector, remove } from './dom/element'
 import { flushPostFlushCbs, queuePostRenderEffect } from './scheduler'
+import { proxyRefs } from '@vue/reactivity'
 import { invokeLifecycle } from './componentLifecycle'
 import { VaporLifecycleHooks } from './apiLifecycle'
 
@@ -11,6 +13,43 @@ export type Fragment = {
   nodes: Block
   anchor?: Node
   [fragmentKey]: true
+}
+
+export function setupComponent(instance: ComponentInternalInstance): void {
+  const reset = setCurrentInstance(instance)
+  instance.scope.run(() => {
+    const { component, props, emit, attrs } = instance
+    const ctx = { expose: () => {}, emit, attrs }
+
+    const setupFn = isFunction(component) ? component : component.setup
+    const stateOrNode = setupFn && setupFn(props, ctx)
+
+    let block: Block | undefined
+
+    if (
+      stateOrNode &&
+      (stateOrNode instanceof Node ||
+        isArray(stateOrNode) ||
+        (stateOrNode as any)[fragmentKey])
+    ) {
+      block = stateOrNode as Block
+    } else if (isObject(stateOrNode)) {
+      instance.setupState = proxyRefs(stateOrNode)
+    }
+    if (!block && component.render) {
+      block = component.render(instance.setupState)
+    }
+
+    if (block instanceof DocumentFragment) {
+      block = Array.from(block.childNodes)
+    }
+    if (!block) {
+      // TODO: warn no template
+      block = []
+    }
+    return (instance.block = block)
+  })
+  reset()
 }
 
 export function render(
