+++ packages/runtime-vapor/src/component.ts
@@ -67,8 +67,6 @@ export interface ComponentInternalInstance {
    */
   inheritAttrs?: boolean
 
-  attrsProxy: Data | null
-
   // lifecycle
   isMounted: boolean
   isUnmounted: boolean
@@ -178,9 +176,6 @@ export function createComponentInstance(
     attrs: EMPTY_OBJ,
     refs: EMPTY_OBJ,
 
-    inheritAttrs: component.inheritAttrs,
-    attrsProxy: null,
-
     // lifecycle
     isMounted: false,
     isUnmounted: false,
@@ -234,6 +229,9 @@ export function createComponentInstance(
      * @internal
      */
     // [VaporLifecycleHooks.SERVER_PREFETCH]: null,
+  }
+  if (component.inheritAttrs != null) {
+    instance.inheritAttrs = component.inheritAttrs
   }
   initProps(instance, rawProps, !isFunction(component))
   instance.emit = emit.bind(null, instance)
